refactor(app): mover navegacion principal fuera de sections
