+++ Web Ui/src/app/navigation/MainMenu.tsx
