+++ Web Ui/src/app/navigation/components/LateralMenu.tsx
